requirements.txt
requests>=2.31.0
pandas>=2.0.0
yfinance>=0.2.36
fredapi>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.32.0
sqlalchemy>=2.0.0
feedparser>=6.0.0
schedule>=1.2.0
psycopg2-binary>=2.9.0
plotly>=5.18.0
numpy>=1.26.0
